refactor: move transform tests
