+++ crates/iceberg/src/spec/transform.rs
@@ -658,512 +658,10 @@ impl<'de> Deserialize<'de> for Transform {
     }
 }
 
+/// An enum representing the result of the adjusted projection.
+/// Either being a single adjusted datum or a set.
 #[derive(Debug)]
 enum AdjustedProjection {
     Single(Datum),
     Set(FnvHashSet<Datum>),
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::spec::datatypes::PrimitiveType::{
-        Binary, Date, Decimal, Fixed, Int, Long, String as StringType, Time, Timestamp,
-        Timestamptz, Uuid,
-    };
-    use crate::spec::datatypes::Type::{Primitive, Struct};
-    use crate::spec::datatypes::{NestedField, StructType, Type};
-    use crate::spec::transform::Transform;
-
-    struct TestParameter {
-        display: String,
-        json: String,
-        dedup_name: String,
-        preserves_order: bool,
-        satisfies_order_of: Vec<(Transform, bool)>,
-        trans_types: Vec<(Type, Option<Type>)>,
-    }
-
-    fn check_transform(trans: Transform, param: TestParameter) {
-        assert_eq!(param.display, format!("{trans}"));
-        assert_eq!(param.json, serde_json::to_string(&trans).unwrap());
-        assert_eq!(trans, serde_json::from_str(param.json.as_str()).unwrap());
-        assert_eq!(param.dedup_name, trans.dedup_name());
-        assert_eq!(param.preserves_order, trans.preserves_order());
-
-        for (other_trans, satisfies_order_of) in param.satisfies_order_of {
-            assert_eq!(
-                satisfies_order_of,
-                trans.satisfies_order_of(&other_trans),
-                "Failed to check satisfies order {}, {}, {}",
-                trans,
-                other_trans,
-                satisfies_order_of
-            );
-        }
-
-        for (input_type, result_type) in param.trans_types {
-            assert_eq!(result_type, trans.result_type(&input_type).ok());
-        }
-    }
-
-    #[test]
-    fn test_bucket_transform() {
-        let trans = Transform::Bucket(8);
-
-        let test_param = TestParameter {
-            display: "bucket[8]".to_string(),
-            json: r#""bucket[8]""#.to_string(),
-            dedup_name: "bucket[8]".to_string(),
-            preserves_order: false,
-            satisfies_order_of: vec![
-                (Transform::Bucket(8), true),
-                (Transform::Bucket(4), false),
-                (Transform::Void, false),
-                (Transform::Day, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), Some(Primitive(Int))),
-                (Primitive(Date), Some(Primitive(Int))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    Some(Primitive(Int)),
-                ),
-                (Primitive(Fixed(8)), Some(Primitive(Int))),
-                (Primitive(Int), Some(Primitive(Int))),
-                (Primitive(Long), Some(Primitive(Int))),
-                (Primitive(StringType), Some(Primitive(Int))),
-                (Primitive(Uuid), Some(Primitive(Int))),
-                (Primitive(Time), Some(Primitive(Int))),
-                (Primitive(Timestamp), Some(Primitive(Int))),
-                (Primitive(Timestamptz), Some(Primitive(Int))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_truncate_transform() {
-        let trans = Transform::Truncate(4);
-
-        let test_param = TestParameter {
-            display: "truncate[4]".to_string(),
-            json: r#""truncate[4]""#.to_string(),
-            dedup_name: "truncate[4]".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Truncate(4), true),
-                (Transform::Truncate(2), false),
-                (Transform::Bucket(4), false),
-                (Transform::Void, false),
-                (Transform::Day, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), Some(Primitive(Binary))),
-                (Primitive(Date), None),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    Some(Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    })),
-                ),
-                (Primitive(Fixed(8)), None),
-                (Primitive(Int), Some(Primitive(Int))),
-                (Primitive(Long), Some(Primitive(Long))),
-                (Primitive(StringType), Some(Primitive(StringType))),
-                (Primitive(Uuid), None),
-                (Primitive(Time), None),
-                (Primitive(Timestamp), None),
-                (Primitive(Timestamptz), None),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_identity_transform() {
-        let trans = Transform::Identity;
-
-        let test_param = TestParameter {
-            display: "identity".to_string(),
-            json: r#""identity""#.to_string(),
-            dedup_name: "identity".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Truncate(4), true),
-                (Transform::Truncate(2), true),
-                (Transform::Bucket(4), false),
-                (Transform::Void, false),
-                (Transform::Day, true),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), Some(Primitive(Binary))),
-                (Primitive(Date), Some(Primitive(Date))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    Some(Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    })),
-                ),
-                (Primitive(Fixed(8)), Some(Primitive(Fixed(8)))),
-                (Primitive(Int), Some(Primitive(Int))),
-                (Primitive(Long), Some(Primitive(Long))),
-                (Primitive(StringType), Some(Primitive(StringType))),
-                (Primitive(Uuid), Some(Primitive(Uuid))),
-                (Primitive(Time), Some(Primitive(Time))),
-                (Primitive(Timestamp), Some(Primitive(Timestamp))),
-                (Primitive(Timestamptz), Some(Primitive(Timestamptz))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_year_transform() {
-        let trans = Transform::Year;
-
-        let test_param = TestParameter {
-            display: "year".to_string(),
-            json: r#""year""#.to_string(),
-            dedup_name: "time".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Year, true),
-                (Transform::Month, false),
-                (Transform::Day, false),
-                (Transform::Hour, false),
-                (Transform::Void, false),
-                (Transform::Identity, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), None),
-                (Primitive(Date), Some(Primitive(Int))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    None,
-                ),
-                (Primitive(Fixed(8)), None),
-                (Primitive(Int), None),
-                (Primitive(Long), None),
-                (Primitive(StringType), None),
-                (Primitive(Uuid), None),
-                (Primitive(Time), None),
-                (Primitive(Timestamp), Some(Primitive(Int))),
-                (Primitive(Timestamptz), Some(Primitive(Int))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_month_transform() {
-        let trans = Transform::Month;
-
-        let test_param = TestParameter {
-            display: "month".to_string(),
-            json: r#""month""#.to_string(),
-            dedup_name: "time".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Year, true),
-                (Transform::Month, true),
-                (Transform::Day, false),
-                (Transform::Hour, false),
-                (Transform::Void, false),
-                (Transform::Identity, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), None),
-                (Primitive(Date), Some(Primitive(Int))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    None,
-                ),
-                (Primitive(Fixed(8)), None),
-                (Primitive(Int), None),
-                (Primitive(Long), None),
-                (Primitive(StringType), None),
-                (Primitive(Uuid), None),
-                (Primitive(Time), None),
-                (Primitive(Timestamp), Some(Primitive(Int))),
-                (Primitive(Timestamptz), Some(Primitive(Int))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_day_transform() {
-        let trans = Transform::Day;
-
-        let test_param = TestParameter {
-            display: "day".to_string(),
-            json: r#""day""#.to_string(),
-            dedup_name: "time".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Year, true),
-                (Transform::Month, true),
-                (Transform::Day, true),
-                (Transform::Hour, false),
-                (Transform::Void, false),
-                (Transform::Identity, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), None),
-                (Primitive(Date), Some(Primitive(Int))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    None,
-                ),
-                (Primitive(Fixed(8)), None),
-                (Primitive(Int), None),
-                (Primitive(Long), None),
-                (Primitive(StringType), None),
-                (Primitive(Uuid), None),
-                (Primitive(Time), None),
-                (Primitive(Timestamp), Some(Primitive(Int))),
-                (Primitive(Timestamptz), Some(Primitive(Int))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_hour_transform() {
-        let trans = Transform::Hour;
-
-        let test_param = TestParameter {
-            display: "hour".to_string(),
-            json: r#""hour""#.to_string(),
-            dedup_name: "time".to_string(),
-            preserves_order: true,
-            satisfies_order_of: vec![
-                (Transform::Year, true),
-                (Transform::Month, true),
-                (Transform::Day, true),
-                (Transform::Hour, true),
-                (Transform::Void, false),
-                (Transform::Identity, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), None),
-                (Primitive(Date), None),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    None,
-                ),
-                (Primitive(Fixed(8)), None),
-                (Primitive(Int), None),
-                (Primitive(Long), None),
-                (Primitive(StringType), None),
-                (Primitive(Uuid), None),
-                (Primitive(Time), None),
-                (Primitive(Timestamp), Some(Primitive(Int))),
-                (Primitive(Timestamptz), Some(Primitive(Int))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    None,
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_void_transform() {
-        let trans = Transform::Void;
-
-        let test_param = TestParameter {
-            display: "void".to_string(),
-            json: r#""void""#.to_string(),
-            dedup_name: "void".to_string(),
-            preserves_order: false,
-            satisfies_order_of: vec![
-                (Transform::Year, false),
-                (Transform::Month, false),
-                (Transform::Day, false),
-                (Transform::Hour, false),
-                (Transform::Void, true),
-                (Transform::Identity, false),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), Some(Primitive(Binary))),
-                (Primitive(Date), Some(Primitive(Date))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    Some(Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    })),
-                ),
-                (Primitive(Fixed(8)), Some(Primitive(Fixed(8)))),
-                (Primitive(Int), Some(Primitive(Int))),
-                (Primitive(Long), Some(Primitive(Long))),
-                (Primitive(StringType), Some(Primitive(StringType))),
-                (Primitive(Uuid), Some(Primitive(Uuid))),
-                (Primitive(Time), Some(Primitive(Time))),
-                (Primitive(Timestamp), Some(Primitive(Timestamp))),
-                (Primitive(Timestamptz), Some(Primitive(Timestamptz))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    Some(Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()]))),
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
-
-    #[test]
-    fn test_known_transform() {
-        let trans = Transform::Unknown;
-
-        let test_param = TestParameter {
-            display: "unknown".to_string(),
-            json: r#""unknown""#.to_string(),
-            dedup_name: "unknown".to_string(),
-            preserves_order: false,
-            satisfies_order_of: vec![
-                (Transform::Year, false),
-                (Transform::Month, false),
-                (Transform::Day, false),
-                (Transform::Hour, false),
-                (Transform::Void, false),
-                (Transform::Identity, false),
-                (Transform::Unknown, true),
-            ],
-            trans_types: vec![
-                (Primitive(Binary), Some(Primitive(StringType))),
-                (Primitive(Date), Some(Primitive(StringType))),
-                (
-                    Primitive(Decimal {
-                        precision: 8,
-                        scale: 5,
-                    }),
-                    Some(Primitive(StringType)),
-                ),
-                (Primitive(Fixed(8)), Some(Primitive(StringType))),
-                (Primitive(Int), Some(Primitive(StringType))),
-                (Primitive(Long), Some(Primitive(StringType))),
-                (Primitive(StringType), Some(Primitive(StringType))),
-                (Primitive(Uuid), Some(Primitive(StringType))),
-                (Primitive(Time), Some(Primitive(StringType))),
-                (Primitive(Timestamp), Some(Primitive(StringType))),
-                (Primitive(Timestamptz), Some(Primitive(StringType))),
-                (
-                    Struct(StructType::new(vec![NestedField::optional(
-                        1,
-                        "a",
-                        Primitive(Timestamp),
-                    )
-                    .into()])),
-                    Some(Primitive(StringType)),
-                ),
-            ],
-        };
-
-        check_transform(trans, test_param);
-    }
 }
